+++ src/main/kotlin/recharts/component/label/LabelPositionObject.kt
@@ -1,6 +1,6 @@
 package recharts.component.label
 
-external interface LabelPositionPosition {
+external interface LabelPositionObject {
     var x: Number?
     var y: Number?
 }