+++ src/main/kotlin/recharts/kotlin/properties/LabelContentTypeOrNullProp.kt
@@ -1,22 +1,22 @@
 package recharts.kotlin.properties
 
-import recharts.component.label.LabelContentType
+import recharts.component.label.ContentType
 import recharts.component.label.LabelProps
 import kotlin.properties.ReadWriteProperty
 import kotlin.reflect.KProperty
 
-class LabelContentTypeOrNullProp : ReadWriteProperty<LabelProps, LabelContentType?> {
-    override fun getValue(thisRef: LabelProps, property: KProperty<*>): LabelContentType? =
+class LabelContentTypeOrNullProp : ReadWriteProperty<LabelProps, ContentType?> {
+    override fun getValue(thisRef: LabelProps, property: KProperty<*>): ContentType? =
         when (val value = thisRef.asDynamic()[property.name]) {
             //is (props: Props) -> ReactNode -> LabelContentType.Function(value as (props: Props) -> ReactNode)
             //is ReactElement -> LabelContentType.ReactElement(value as ReactElement)
             else -> null
         }
 
-    override fun setValue(thisRef: LabelProps, property: KProperty<*>, value: LabelContentType?) {
+    override fun setValue(thisRef: LabelProps, property: KProperty<*>, value: ContentType?) {
         thisRef.asDynamic()[property.name] = when (value) {
             //is LabelContentType.Function -> value.value
-            is LabelContentType.ReactElement -> value.value
+            is ContentType.ReactElement -> value.value
             null -> null
         }
     }